🚚 chore(map3d): Move files to fit better in project structure
