+++ src/components/Map3D/createLineVertex.ts
