+++ src/typescript/mapInterfaces.ts
